# src/main/resources/openapi/openapi-geography.yaml
openapi: 3.0.3
info:
  title: Vacapp Geographic Control API
  description: API REST para gestión jerárquica de terrenos ganaderos (Rancho → Sección → Potrero)
  version: 1.0.0
  contact:
    name: Vacapp Team
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/geography/ranchos:
    post:
      tags:
        - Ranchos
      summary: Crear nuevo rancho
      operationId: createRancho
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRanchoRequest'
      responses:
        '201':
          description: Rancho creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RanchoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
    
    get:
      tags:
        - Ranchos
      summary: Listar ranchos del tenant
      operationId: listRanchos
      parameters:
        - name: page
          in: query
          description: Número de página (0-based)
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Tamaño de página
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Lista de ranchos
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RanchoResponse'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/geography/ranchos/{id}:
    get:
      tags:
        - Ranchos
      summary: Obtener rancho por ID
      operationId: getRanchoById
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Rancho encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RanchoResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      tags:
        - Ranchos
      summary: Actualizar rancho
      operationId: updateRancho
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRanchoRequest'
      responses:
        '200':
          description: Rancho actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RanchoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      tags:
        - Ranchos
      summary: Archivar rancho
      operationId: archiveRancho
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '204':
          description: Rancho archivado exitosamente
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      description: UUID del recurso
      schema:
        type: string
        format: uuid
  schemas:
...
    CreateRanchoRequest:
      type: object
      required:
        - nombre
        - superficieTotal
      properties:
        nombre:
          type: string
          minLength: 2
          maxLength: 100
          description: Nombre del rancho
          example: "Rancho El Encino"
        superficieTotal:
          type: number
          format: decimal
          minimum: 0.01
          maximum: 999999999
          description: Superficie total en metros cuadrados
          example: 50000.50
        descripcion:
          type: string
          maxLength: 500
          description: Descripción opcional del rancho
          example: "Rancho principal con zona de pastoreo norte"
    UpdateRanchoRequest:
      type: object
      required:
        - nombre
        - superficieTotal
      properties:
        nombre:
          type: string
          minLength: 2
          maxLength: 100
          example: "Rancho El Encino Actualizado"
        superficieTotal:
          type: number
          format: decimal
          minimum: 0.01
          maximum: 999999999
          example: 55000.00
        descripcion:
          type: string
          maxLength: 500
    
    # ===== REQUEST DTOs - SECCION =====
...
    RanchoResponse:
      type: object
      properties:
        ranchoId:
          type: string
          format: uuid
        nombre:
          type: string
        superficieTotal:
          type: number
          format: decimal
        superficieDisponible:
          type: number
          format: decimal
        superficieUsada:
          type: number
          format: decimal
        porcentajeUso:
          type: number
          format: decimal
          description: Porcentaje de uso (0.00 - 100.00)
        descripcion:
          type: string
        status:
          type: string
          enum: [ACTIVE, ARCHIVED]
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 0
          description: Página actual (0-based)
        size:
          type: integer
          minimum: 1
          description: Tamaño de página
        total:
          type: integer
          minimum: 0
          description: Total de elementos
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    BadRequest:
      description: Petición inválida (validación fallida)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Autenticación requerida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflicto (ej. nombre duplicado)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'